requirements.txt
streamlit
transformers
torch
Pillow
watchdog
rembg
onnxruntime
numpy<2.3
